requirements.txt
AWSIoTPythonSDK
python-dotenv